requirements.txt
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
Werkzeug==2.0.2
